chore(rest): add missing test for process definition delegation
related to #OPT-33

+++ backend/src/main/java/org/camunda/optimize/service/es/ProcessDefinitionReader.java
@@ -41,7 +41,7 @@ public class ProcessDefinitionReader {
       .setQuery(query)
       .get();
 
-    int numberOfProcessDefinitions = (int) (int) sr.getHits().totalHits();
+    int numberOfProcessDefinitions = (int) sr.getHits().totalHits();
     List<ProcessDefinitionDto> list = new ArrayList<>(numberOfProcessDefinitions);
     for (SearchHit hit : sr.getHits().getHits()) {
       String content = hit.getSourceAsString();
